Shoot

+++ src/main/java/com/mycompany/scannonpreview/objects/Vector2D.java
@@ -9,10 +9,10 @@ package com.mycompany.scannonpreview.objects;
  */
 public class Vector2D {
 
-    public int x = 0;
-    public int y = 0;
+    public double x = 0;
+    public double y = 0;
 
-    public Vector2D(int x, int y) {
+    public Vector2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
